23-04-11

+++ chapter06_1/src/main/java/com/bean/SungJukDTO.java
@@ -11,5 +11,5 @@ public class SungJukDTO {
 	private int eng;
 	private int math;
 	private int tot;
-	private int avg;
+	private double avg;
 }
